+++ Services/Interfaces/ICourseSchedule.cs
@@ -10,13 +10,13 @@ namespace EduReg.Services.Interfaces
 
       //  Task<GeneralResponse> CreateCourseScheduleAsync(byte[] model); // EXCEL, CSV, Flat file upload
 
-        Task<GeneralResponse> UpdateCourseScheduleAsync(int Id, CourseScheduleDto model);
+        Task<GeneralResponse> UpdateCourseScheduleAsync(long Id, CourseScheduleDto model);
 
-        Task<GeneralResponse> DeleteCourseScheduleAsync(int Id);
+        Task<GeneralResponse> DeleteCourseScheduleAsync(long Id);
         Task<GeneralResponse> DeleteManyCourseSchedulesAsync(List<CourseScheduleDto> model);
-        Task<GeneralResponse> DeleteManyCourseSchedulesAsync(List<int> Id);
+        Task<GeneralResponse> DeleteManyCourseSchedulesAsync(List<long> Id);
 
-        Task<GeneralResponse> GetCourseScheduleByIdAsync(int Id);
+        Task<GeneralResponse> GetCourseScheduleByIdAsync(long Id);
         Task<GeneralResponse> GetCourseScheduleByCourseCodeAsync(string courseCode);
         Task<GeneralResponse> GetAllCourseSchedulesAsync();
