+++ Services/Interfaces/IDepartments.cs
@@ -6,9 +6,9 @@ namespace EduReg.Services.Interfaces
     public interface IDepartments
     {
         Task<GeneralResponse> CreateDepartmentAsync(DepartmentsDto model);
-        Task<GeneralResponse> UpdateDepartmentAsync(int Id, DepartmentsDto model);
-        Task<GeneralResponse> DeleteDepartmentAsync(int Id);
-        Task<GeneralResponse> GetDepartmentByIdAsync(int Id);
+        Task<GeneralResponse> UpdateDepartmentAsync(long Id, DepartmentsDto model);
+        Task<GeneralResponse> DeleteDepartmentAsync(long Id);
+        Task<GeneralResponse> GetDepartmentByIdAsync(long Id);
         Task<GeneralResponse> GetDepartmentByNameAsync(string DepartmentName);
         Task<GeneralResponse> GetAllDepartmentsAsync();
     }
