+++ Services/Interfaces/IProgrammes.cs
@@ -6,9 +6,9 @@ namespace EduReg.Services.Interfaces
     public interface IProgrammes
     {
         Task<GeneralResponse> CreateProgrammeAsync(ProgrammesDto model);
-        Task<GeneralResponse> UpdateProgrammeAsync(int Id, ProgrammesDto model);
-        Task<GeneralResponse> DeleteProgrammeAsync(int Id);
-        Task<GeneralResponse> GetProgrammeByIdAsync(int Id);
+        Task<GeneralResponse> UpdateProgrammeAsync(long Id, ProgrammesDto model);
+        Task<GeneralResponse> DeleteProgrammeAsync(long Id);
+        Task<GeneralResponse> GetProgrammeByIdAsync(long Id);
         Task<GeneralResponse> GetProgrammeByNameAsync(string ProgrammeName);
         Task<GeneralResponse> GetAllProgrammesAsync();
     }
