+++ Services/Interfaces/IStudentFeePaymentService.cs
@@ -4,7 +4,7 @@ namespace EduReg.Services.Interfaces
 {
     public interface IStudentFeePaymentService
     {
-        Task<GeneralResponse> UpdateStudentPaymentAsync(string matricNumber, List<int> paidItemIds);
+        Task<GeneralResponse> UpdateStudentPaymentAsync(string matricNumber, List<long> paidItemIds);
 
 
     }
